+++ Utils/Synth.hpp
@@ -5,7 +5,7 @@
 class Synth
 {
 private:
-    static float CalculateSample(float time, float frequency, float sampleRate);
+    static float CalculateSample(int time, float frequency, float sampleRate);
     
 public:
     static SampleList Synthesize(float frequency, float duration, float sampleRate);
